Refactor physics system and introduce Space class
- Add Space class to represent the physical environment
- Introduce Fizicks class as a wrapper for physics simulation
- Remove individual law classes (FirstLaw, SecondLaw, ThirdLaw)
- Update Motion class to handle physics updates
- Improve type hinting and documentation
