+++ fizicks/__init__.py
@@ -1,1 +1,2 @@
-from fizicks.main import *
+from fizicks.main import Fizicks
+from fizicks.data import *
